Differential Revision: [D36377117](https://our.internmc.facebook.com/intern/diff/D36377117)

This patch is about enabling mkldnn on float32 and bfloat16 for transposed convolution on CPU path.

cc jgong5 @XiaobingSuper sanchitintel ashokei jingxu10

[ghstack-poisoned]

+++ .circleci/scripts/binary_ios_test.sh
@@ -1,30 +1,19 @@
 #!/bin/bash
 set -ex -o pipefail
 
+if ! [ "$IOS_PLATFORM" == "SIMULATOR" ]; then
+    exit 0
+fi
+
 echo ""
 echo "DIR: $(pwd)"
 PROJ_ROOT=/Users/distiller/project
 cd ${PROJ_ROOT}/ios/TestApp
 # install fastlane
 sudo gem install bundler && bundle install
-# install certificates
-echo "${IOS_CERT_KEY_2022}" >> cert.txt
-base64 --decode cert.txt -o Certificates.p12
-rm cert.txt
-bundle exec fastlane install_root_cert
-bundle exec fastlane install_dev_cert
-# install the provisioning profile
-PROFILE=PyTorch_CI_2022.mobileprovision
-PROVISIONING_PROFILES=~/Library/MobileDevice/Provisioning\ Profiles
-mkdir -pv "${PROVISIONING_PROFILES}"
-cd "${PROVISIONING_PROFILES}"
-echo "${IOS_SIGN_KEY_2022}" >> cert.txt
-base64 --decode cert.txt -o ${PROFILE}
-rm cert.txt
 # run the ruby build script
 if ! [ -x "$(command -v xcodebuild)" ]; then
     echo 'Error: xcodebuild is not installed.'
     exit 1
 fi
-PROFILE=PyTorch_CI_2022
-ruby ${PROJ_ROOT}/scripts/xcode_build.rb -i ${PROJ_ROOT}/build_ios/install -x ${PROJ_ROOT}/ios/TestApp/TestApp.xcodeproj -p ${IOS_PLATFORM} -c ${PROFILE} -t ${IOS_DEV_TEAM_ID}
+ruby ${PROJ_ROOT}/scripts/xcode_build.rb -i ${PROJ_ROOT}/build_ios/install -x ${PROJ_ROOT}/ios/TestApp/TestApp.xcodeproj -p ${IOS_PLATFORM}
